Resolve merge conflicts:
- `YTsaurusDictionarySource.cpp`: Use master's `nullptr` init + named collection logic,
  fix typo `YTsarususDictionarySource` -> `YTsaurusDictionarySource`, adopt
  `VectorWithMemoryTracking` parameter types from master
- `StorageYTsaurus.cpp`: Pass `&args.table_id` to `getConfiguration` as master requires
- `TableFunctionYtsaurus.cpp`: Keep PR's `check_table_schema` setting propagation,
  add master's `ASTSetQuery` parsing and named collection (2-arg) support

+++ src/Common/StudentTTest.h
@@ -36,7 +36,7 @@ private:
 
         double var() const
         {
-            return (squares_sum - (sum * sum / size)) / static_cast<double>(size - 1);
+            return (squares_sum - (sum * sum / static_cast<double>(size))) / static_cast<double>(size - 1);
         }
 
         void clear()
